+++ SlaughterHouse/SlaughterHouse/Entities/AnimalPart.cs
@@ -4,6 +4,6 @@ public class AnimalPart
 {
     public int AnimalId { get; set; }
     public int partID  { get; set; }
-    public int partWeight { get; set; }
+    public decimal partWeight { get; set; }
     public PartType partType { get; set; }
 }
