+++ SlaughterHouse/SlaughterHouse/Entities/Product.cs
@@ -5,5 +5,5 @@ public class Product
     public int productId { get; set; }
     public int trayId { get; set; }
     public ProductType productType { get; set; }
-    public int productWeight { get; set; }
+    public decimal productWeight { get; set; }
 }
