+++ SlaughterHouse/SlaughterHouse/Entities/Tray.cs
@@ -5,7 +5,7 @@ public class Tray
     public int TrayId { get; set; }
     public PartType partType { get; set; }
     public int maxCapacity { get; set; }
-    public int currentWeight { get; set; }
+    public decimal currentWeight { get; set; }
 
     public Tray(int trayId, PartType partType, int maxCapacity)
     {
